pyproject.toml
[tool.poetry]
name = "hierarchical_data"
version = "0.1.0"
description = ""
authors = ["peyton glover <peyton.vx@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
